Refactor entrypoints. Remove old macro
