+++ algorithmia/Cargo.toml
@@ -3,7 +3,6 @@ name = "algorithmia"
 version = "2.1.1"
 license = "MIT"
 authors = ["Anthony Nowell <anthony@algorithmia.com>"]
-
 description = "Client for calling algorithms and managing data hosted on the Algorithmia marketplace"
 documentation = "http://docs.rs/algorithmia"
 repository = "https://github.com/algorithmiaio/algorithmia-rust"
@@ -22,10 +21,10 @@ url = "1.5"
 serde = "1.0"
 serde_json = "1.0"
 serde_derive = "1.0"
-base64 = "0.6.0"
+base64 = "0.9.0"
 error-chain = "0.8.1"
 mime = "0.3.3"
-reqwest = "0.7"
+reqwest = "0.8"
 chrono = { version = "0.4", features = ["serde"] }
 
 [dependencies.hyper]
@@ -40,6 +39,6 @@ optional = true
 rustc_version = "0.2.1"
 
 [features]
+nightly = []
+entrypoint = ["algorithmia-entrypoint"]
 default = []
-entrypoint = []
-nightly = ["entrypoint", "algorithmia-entrypoint"]
